+++ src/common-lib/error-reporter/errorReporter.test.ts
@@ -65,27 +65,38 @@ describe("common-lib/error-reporter", () => {
     it("returns false if the ua string doesn't contain words in the disallow list", () => {
       expect(
         matchesUserAgent(
-          "Mozilla/5.0 (Macintosh; Intel Mac OS X 11_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4619.141 Safari/537.36"
-        )
+          "Mozilla/5.0 (Macintosh; Intel Mac OS X 11_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4619.141 Safari/537.36",
+        ),
       ).toBe(false);
     });
     it("returns true if ua string contains 'percy'", () => {
       expect(
         matchesUserAgent(
-          "Mozilla/5.0 (Macintosh; Intel Mac OS X 11_4) percy AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4619.141 Safari/537.36"
-        )
+          "Mozilla/5.0 (Macintosh; Intel Mac OS X 11_4) percy AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4619.141 Safari/537.36",
+        ),
       ).toBe(true);
     });
   });
   describe("matchesIgnoredError", () => {
     it("returns false if the error should not be ignored", () => {
-      const shouldIgnore = matchesIgnoredError(
-        "Proper error that should be reported"
-      );
+      const shouldIgnore = matchesIgnoredError({
+        errorMessage: "Proper error that should be reported",
+        stacktrace: [],
+      });
       expect(shouldIgnore).toBe(false);
     });
-    it("returns true if the error should be ignored", () => {
-      const shouldIgnore = matchesIgnoredError("Test error");
+    it("returns true if the error should be ignored based on message", () => {
+      const shouldIgnore = matchesIgnoredError({
+        errorMessage: "Test error",
+        stacktrace: [],
+      });
+      expect(shouldIgnore).toBe(true);
+    });
+    it("returns true if the error should be ignored based on stacktrace", () => {
+      const shouldIgnore = matchesIgnoredError({
+        errorMessage: "Proper error message",
+        stacktrace: [{ file: "https://OAK_TEST_ERROR_STACKTRACE_FILE.js" }],
+      });
       expect(shouldIgnore).toBe(true);
     });
   });
@@ -96,6 +107,7 @@ describe("common-lib/error-reporter", () => {
         errors: [
           {
             errorMessage: "real error",
+            stacktrace: [{ file: "real file" }],
           },
         ],
       } as BugsnagEvent;
@@ -115,9 +127,10 @@ describe("common-lib/error-reporter", () => {
         errors: [
           {
             errorMessage: "Test error",
+            stacktrace: [],
           },
         ],
-      } as BugsnagEvent;
+      } as unknown as BugsnagEvent;
       const result = getBugsnagOnError({ logger })(event);
       expect(result).toBe(false);
     });
@@ -163,7 +176,7 @@ describe("common-lib/error-reporter", () => {
 
       await reportError("test thing");
       expect(consoleLog).toHaveBeenCalledWith(
-        "Failed to send error to bugsnag:"
+        "Failed to send error to bugsnag:",
       );
       expect(consoleError).toHaveBeenCalledWith("bad thing");
       expect(consoleLog).toHaveBeenCalledWith("Original error:");
@@ -171,7 +184,7 @@ describe("common-lib/error-reporter", () => {
     });
     test("adds originalError if error is OakError", () => {
       const originalError = new Error(
-        "some error from somewhere (not our fault!)"
+        "some error from somewhere (not our fault!)",
       );
       const oakError = new OakError({ code: "misc/unknown", originalError });
       reportError(oakError);
@@ -216,7 +229,7 @@ describe("common-lib/error-reporter", () => {
       reportError(error);
       expect(mockNotify).toHaveBeenCalledTimes(1);
       expect(consoleWarn).toHaveBeenCalledWith(
-        expect.stringContaining("already reported")
+        expect.stringContaining("already reported"),
       );
     });
   });