+++ src/common-lib/error-reporter/errorReporter.ts
@@ -21,7 +21,17 @@ export const matchesUserAgent = (ua: string) => {
   return userAgentsToMatch.some((regex) => regex.test(ua));
 };
 
-export const matchesIgnoredError = (message: string) => {
+export const matchesIgnoredError = (error: {
+  errorMessage: string;
+  stacktrace: { file: string }[];
+}) => {
+  const filesToMatch = [
+    // Testing
+    /OAK_TEST_ERROR_STACKTRACE_FILE/i,
+    // Don't error external Hubspot script problems
+    /\/\/js\.hubspot\.com/i,
+    /\/\/js\.hsleadflows\.net/i,
+  ];
   const messagesToMatch = [
     // Testing
     /Test error/i,
@@ -29,12 +39,18 @@ export const matchesIgnoredError = (message: string) => {
     // https://github.com/oaknational/Oak-Web-Application/issues/999
     /Multiple lead flow scripts are trying to run on the current page/i,
     /t.report is not a function/i,
+    /null is not an object (evaluating 'e.portalId')/i,
+    /Hubspot script failed to load/i,
   ];
-  return messagesToMatch.some((regex) => regex.test(message));
+  return (
+    filesToMatch.some((regex) =>
+      error.stacktrace.some((stacktrace) => regex.test(stacktrace.file)),
+    ) || messagesToMatch.some((regex) => regex.test(error.errorMessage))
+  );
 };
 
 export function getBugsnagOnError(
-  { logger }: { logger: Logger } = { logger: console }
+  { logger }: { logger: Logger } = { logger: console },
 ) {
   return function bugsnagOnError(event: Event) {
     const { userAgent } = event.device;
@@ -49,10 +65,9 @@ export function getBugsnagOnError(
     // Ignore some known errors that aren't user impacting but do mess up the stability metrics.
     const firstError = event?.errors[0];
     if (firstError !== undefined) {
-      const errorMessage = firstError.errorMessage;
-      const shouldIgnore = matchesIgnoredError(errorMessage);
+      const shouldIgnore = matchesIgnoredError(firstError);
       if (shouldIgnore) {
-        logger.warn(`Ignoring known issue: ${errorMessage}`);
+        logger.warn(`Ignoring known issue: ${firstError.errorMessage}`);
         return false;
       }
     }
@@ -138,7 +153,7 @@ const errorify = (maybeError: unknown): Error => {
     return new Error(message);
   } catch (jsonStringifyError) {
     return new Error(
-      `Failed to stringify maybeError, type: ${typeof maybeError}`
+      `Failed to stringify maybeError, type: ${typeof maybeError}`,
     );
   }
 };
@@ -181,7 +196,7 @@ type Logger = Pick<typeof console, "log" | "warn" | "error">;
 const errorReporter = (
   context: string,
   metadata?: Record<string, unknown>,
-  { logger }: { logger: Logger } = { logger: console }
+  { logger }: { logger: Logger } = { logger: console },
 ) => {
   const reportError = async (maybeError: MaybeError, data?: ErrorData) => {
     try {