+++ src/components/NewButton.tsx/NewButtonLabelWithScreenReaderTitle.tsx
@@ -1,7 +1,6 @@
 import { FC } from "react";
 import styled from "styled-components";
 
-import ellipsis from "../../styles/ellipsis";
 import typography from "../../styles/utils/typography";
 import { Span } from "../Typography";
 import ScreenReaderOnly from "../ScreenReaderOnly/ScreenReaderOnly";
@@ -25,13 +24,12 @@ const ButtonLabelWithScreenReaderTitle: FC<ButtonLabelProps> = (props) => {
 };
 
 export const StyledButtonLabelWithScreenReaderTitle = styled(
-  ButtonLabelWithScreenReaderTitle
+  ButtonLabelWithScreenReaderTitle,
 )`
   display: inline-block;
   vertical-align: text-top;
   max-width: 100%;
   ${typography}
-  ${ellipsis}
 `;
 
 StyledButtonLabelWithScreenReaderTitle.defaultProps = {